Rename file integration test to blob
