+++ tests/integration/blob/blob.js
