+++ tests/integration/handlers.js
@@ -1,6 +1,6 @@
+export { handler as blob } from './blob/blob.js';
 export { handler as btoa } from './btoa/btoa.js';
 export { handler as performance } from './performance/performance.js';
 export { handler as crypto } from './crypto/crypto.js';
 export { handler as timers } from './timers/timers.js';
 export { handler as fetch } from './fetch/fetch.js';
-export { handler as file } from './file/file.js';
