+++ tests/tests.cmake
@@ -42,9 +42,9 @@ test_e2e(tla)
 test_e2e(stream-forwarding)
 test_e2e(multi-stream-forwarding)
 
+test_integration(blob)
 test_integration(btoa)
 test_integration(crypto)
 test_integration(fetch)
-test_integration(file)
 test_integration(performance)
 test_integration(timers)
